Moved ContainerProviders into a dedicated package
