+++ src/main/java/com/drakmyth/minecraft/manufactory/blocks/GrinderBlock.java
@@ -6,11 +6,11 @@
 package com.drakmyth.minecraft.manufactory.blocks;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
-import com.drakmyth.minecraft.manufactory.menus.GrinderContainerProvider;
-import com.drakmyth.minecraft.manufactory.menus.GrinderUpgradeContainerProvider;
 import com.drakmyth.minecraft.manufactory.init.ModTags;
 import com.drakmyth.minecraft.manufactory.init.ModTileEntityTypes;
 import com.drakmyth.minecraft.manufactory.items.upgrades.IPowerUpgrade;
+import com.drakmyth.minecraft.manufactory.menus.providers.GrinderContainerProvider;
+import com.drakmyth.minecraft.manufactory.menus.providers.GrinderUpgradeContainerProvider;
 import com.drakmyth.minecraft.manufactory.network.ModPacketHandler;
 import com.drakmyth.minecraft.manufactory.network.OpenContainerWithUpgradesPacket;
 import com.drakmyth.minecraft.manufactory.power.IPowerBlock;
