+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/providers/RockDrillUpgradeContainerProvider.java
@@ -3,9 +3,10 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.menus;
+package com.drakmyth.minecraft.manufactory.menus.providers;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
+import com.drakmyth.minecraft.manufactory.menus.RockDrillUpgradeMenu;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -13,28 +14,28 @@ import org.apache.logging.log4j.Logger;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.MenuProvider;
-import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class GrinderUpgradeContainerProvider implements MenuProvider {
+public class RockDrillUpgradeContainerProvider implements MenuProvider {
     private static final Logger LOGGER = LogManager.getLogger();
-    private BlockPos pos;
+    private ItemStack stack;
 
-    public GrinderUpgradeContainerProvider(BlockPos pos) {
-        this.pos = pos;
+    public RockDrillUpgradeContainerProvider(ItemStack stack) {
+        this.stack = stack;
     }
 
     @Override
     public AbstractContainerMenu createMenu(int windowId, Inventory playerInventory, Player player) {
-        LOGGER.debug(LogMarkers.CONTAINER, "Creating Grinder upgrade gui...");
-        return new GrinderUpgradeMenu(windowId, new InvWrapper(playerInventory), player, pos);
+        LOGGER.debug(LogMarkers.CONTAINER, "Creating Rock Drill upgrade gui...");
+        return new RockDrillUpgradeMenu(windowId, new InvWrapper(playerInventory), player, stack);
     }
 
     @Override
     public Component getDisplayName() {
-        return new TextComponent("Grinder");
+        return new TextComponent("Rock Drill");
     }
 }
